requirements.txt
#asyncio>=3.4.3
#requests==2.23.0
#botbuilder-core>=4.14.0
#botframework-connector>=4.14.0
#botbuilder-schema>=4.14.0
#botbuilder-dialogs>=4.14.0
#botbuilder-ai>=4.14.0
#botbuilder-applicationinsights>=4.14.0
#botbuilder-integration-applicationinsights-aiohttp>=4.14.0
#datatypes-date-time>=1.0.0.a1
#azure-cognitiveservices-language-luis==0.2.0
#msrest>=0.6.10
#multidict >= 4.5.2


aiohttp==3.7.4.post0
#aiounittest==1.3.0
#applicationinsights==0.11.9
#asgiref==3.6.0
#async-timeout==3.0.1
#asyncio==3.4.3
#attrs==22.2.0
azure-cognitiveservices-language-luis==0.2.0
#azure-common==1.1.28
botbuilder-ai==4.14.3
botbuilder-applicationinsights==4.14.3
botbuilder-core==4.14.3
botbuilder-dialogs==4.14.3
botbuilder-integration-applicationinsights-aiohttp==4.14.3
botbuilder-schema==4.14.3
botframework-connector==4.14.3
botframework-streaming==4.14.3
datatypes-date-time==1.0.0a2
datedelta==1.4
dateparser==1.1.7
msrest==0.6.21
msrestazure==0.6.4
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.23.0